+++ src/main/java/racingcar/domain/racing/RepeatCount.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 public class RepeatCount {
